Start working on GUI.
Add css and DOM elements
There is a left side of webpage and right side. Left is the canvas (and picture) while the right is a settings panel.
It is comprised of a working download button, and a currently useless select element (dropdown). It has all the styles of image (eg greyscale, inverse colors)